update:float型化

+++ Assets/Scripts/Components/EnemyComponent.cs
@@ -4,9 +4,9 @@ using UnityEngine.UI;
 public class EnemyComponent : MonoBehaviour
 {
     public float speed;
-    public int maxHp;
-    public int hp;
-    public int attack;
+    public float maxHp;
+    public float hp;
+    public float attack;
     public float coolDownTimer;
     public float coolTime;
     public GameObject emptyObj;
